Add multi-architecture builds.
Adds support for aarch64 builds.

This is done by adding a new docker file, Dockerfile.multiarch, which
cross-compiles to a given architecture using the correct toolchain for
the machine type.

The original Dockerfile for "native" "amd64 on amd64" builds is kept,
but the variables for library versions are moved to a common location to
prevent repetition when incrementing these.

A build script is added to aid in building and pushing each image and to
also manage the manifest layers.

Because the cross-compiled images are compiling binaries that they
cannot execute, the tests for OpenSSL and PCRE2 are currently skipped.
These could be run in qemu, but the OpenSSL tests seem unreliable in
that environment.