Support aarch64 as a builder
Images can now be compiled either by an x86 or aarch64 builder.

When the builder arch matches the target arch, the image is produced
natively (rather than running through emulation). When the builder arch
and the target arch differ, the appropriate toolchain is installed and
used for image compilation, thus avoiding emulation and slow compile
times for all cases.

Until this change, the builder was always assumed to be x86-based.